Refactor preset to emit an error if updating project fails
