+++ src/constants.ts
@@ -22,3 +22,4 @@ export type UpdateProjectPayload = {
 };
 
 export const PROJECT_UPDATED = `${ADDON_ID}/projectUpdated`;
+export const PROJECT_UPDATING_FAILED = `${ADDON_ID}/projectUpdatingFailed`;
